refactor: split out validation function
